requirements.txt
ultralytics>=8.3.127
Pillow>=9.5
